Completion of exercise 4
Chaining selectors to be able to go deeper into the class.
Doing this can select a child of the class element and allow you to be
more specific with what you are selecting.

Doing so means you can apply a style to multiple elements and have the
child of these be different.